Shrink coolant calibration tables (#1431)
* Use shared function for 2d table printing.

* Bug fix: must load calibrations before call to doUpdates()!!

* Chnge coolant & IAT calibration values to uint8_t, which is our internal temperature type.

* Save Flash: compile updates.cpp as -Os

* Move storage fake to separate file

* Unit test updateTableU16toU8

* Unit test upgradeV25toV26()

* Clean up print_eeprom_layout() output

+++ speeduino/comms.cpp
@@ -445,7 +445,7 @@ static uint8_t toTemperature(byte lo, byte hi)
  * @param values The table values
  * @param bins The table bin values
  */
-static void processTemperatureCalibrationTableUpdate(uint16_t calibrationLength, SensorCalibrationTable calibrationPage, uint16_t *values, uint16_t *bins)
+static void processTemperatureCalibrationTableUpdate(uint16_t calibrationLength, SensorCalibrationTable calibrationPage, uint8_t *values, uint16_t *bins)
 {
   //Temperature calibrations are sent as 32 16-bit values
   if(calibrationLength == 64U)
